Add router module, render page on popstate
