+++ src/components/page-sign-in/page-sign-in.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { Component } from '../../component';
-import { loadPath } from '../..';
+import { loadPath } from '../../modules/router';
 import { Button } from '../button/button';
 import { FormSignIn } from '../form-sign-in/form-sign-in';
