+++ src/components/page-sign-up/page-sign-up.ts
@@ -1,7 +1,7 @@
 'use strict';
 
 import { Component } from '../../component';
-import { loadPath } from '../..';
+import { loadPath } from '../../modules/router';
 import { Button } from '../button/button';
 import { FormSignUp } from '../form-sign-up/form-sign-up';
